Link target issue via its number
Currently, the ticket Title is used as handle for ticket management
(create, update, close), however the title can be changed outside the
scope of he operator (e.g. on Github), so that the operator won't be
able to track it anymore.

This commit allows the operator to "link" with the Github ticket,
storing the ticket Number in the resource's Status. Once the ticket is
linked, the operator can keep track of the same ticket in the future,
regardless the value of the ticket's title.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>